Added delete and list

+++ api/server.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const cors = require("cors");
 const database = require("./database");
-
+const port = 8080;
 
 if (database) {
   const app = express();
@@ -9,7 +9,7 @@ if (database) {
   app.use(cors());
 
   app.post("/images", async (req, res) => {
-    console.log('POST /');
+    console.log("POST /images");
 
     res.set("Access-Control-Allow-Origin", "http://localhost:3000");
     res.set("Access-Control-Allow-Methods", "POST,GET,DELETE, HEAD, OPTIONS");
@@ -19,26 +19,49 @@ if (database) {
 
       if (url !== undefined) {
         await database.createImage(url);
-        res.json({response: "Image url stored in database"});
+        res.json({ StoreResponse: "Image has been inserted in database" });
       }
-    } catch(err) {
-      res.status(500).json({Error: "Something unexepcted happend"});
-      console.log("Oops:", err.message);
+    } catch (err) {
+      res.status(500).json({ Error: "Something unexepcted happend" });
+      console.log("Oops: ", err.message);
     }
-  })
+  });
 
-  /*app.delete("/images/:id", async (req, res) => {
-    const { id } = req.params;
-    await database.deleteImage(id);
-    res.json();
+  app.delete("/images/:id", async (req, res) => {
+    console.log("DELETE /images");
+
+    try {
+      const { id } = req.params;
+      console.log(id);
+
+      if (id !== undefined) {
+        await database.deleteImage(id);
+        res.json({ deleteResponse: "Image delete from database" });
+      } else {
+        res.status(400).json({ Error: "Bad request" });
+      }
+    } catch (err) {
+      res.status(500).json({ Error: "Something unexepcted happend" });
+      console.log("Oops: ", err.message);
+    }
   });
 
   app.get("/images", async (req, res) => {
-    const images = await database.listImages();
-    res.json(images);
-  });*/
+    console.log("GET /images");
 
-  app.listen(8080, () => {
+    try {
+      const images = await database.listImages();
+
+      if (images) {
+        res.json({ imagesResponse: images });
+      }
+    } catch (err) {
+      res.status(500).json({ Error: "Something unexepcted happend" });
+      console.log("Oops: ", err.message);
+    }
+  });
+
+  app.listen(port, () => {
     console.log("app is running on port 8080");
   });
 }
